+++ Main/include/Game.hpp
@@ -29,7 +29,7 @@ struct ScoreReplay
 	int32 currentScore = 0; //< Current score; updated during playback
 	int32 currentMaxScore = 0; //< Current max possible score; updated during playback
 	int32 maxScore = 0;
-	int32 nextHitStat = 0;
+	size_t nextHitStat = 0;
 	Vector<SimpleHitStat> replay;
 };
